requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
cloudpickle==1.6.0
comtypes==1.1.7
cycler==0.10.0
dask==2.30.0
decorator==4.4.2
ecapture==0.1.8
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
more-itertools==8.6.0
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.15.0
pyttsx3==2.90
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
PyYAML==5.3.1
requests==2.25.0
scikit-image==0.14.2
scipy==1.5.4
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tifffile==2020.11.18
toolz==0.11.1
urllib3==1.26.2
wikipedia==1.4.0
wolframalpha==4.1.0
xmltodict==0.12.0
